prettier .ts and .sol

+++ contracts/PoolEth.sol
@@ -5,29 +5,25 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 import "hardhat/console.sol";
 
 contract PoolEth is Ownable {
-
-    enum State {
-        DEPOSITED,
-        WITHDRAWN
-    }
+    enum State { DEPOSITED, WITHDRAWN }
 
     struct Deposit {
-        uint amount;
-        uint date;
+        uint256 amount;
+        uint256 date;
     }
 
     struct Reward {
-        uint idDeposit;
-        uint amount;
-        uint percentage;
+        uint256 idDeposit;
+        uint256 amount;
+        uint256 percentage;
         State state;
     }
 
-    Deposit [] private deposits;
-    mapping(address => Reward []) private rewards; 
-    
-    uint public totalPool;
-    mapping(address => uint) private balances;
+    Deposit[] private deposits;
+    mapping(address => Reward[]) private rewards;
+
+    uint256 public totalPool;
+    mapping(address => uint256) private balances;
 
     address[] private stakers;
     mapping(address => bool) private hasStaked;
@@ -38,21 +34,20 @@ contract PoolEth is Ownable {
     }
 
     function stake() external payable onlyUser {
-
         require(msg.value > 0, "PoolEth: amount can not be 0");
 
         balances[msg.sender] += msg.value;
         totalPool += msg.value;
 
-        if(!hasStaked[msg.sender]) {
+        if (!hasStaked[msg.sender]) {
             stakers.push(msg.sender);
         }
 
         hasStaked[msg.sender] = true;
     }
-    
+
     function compound() external onlyUser {
-        uint rewardAmount = getReward();
+        uint256 rewardAmount = getReward();
         require(rewardAmount > 0, "PoolEth: has no rewards");
         _changeStateRewards();
         balances[msg.sender] += rewardAmount;
@@ -60,25 +55,25 @@ contract PoolEth is Ownable {
     }
 
     function harvest() external onlyUser {
-        uint amount = _availableAmount();
+        uint256 amount = _availableAmount();
         require(amount > 0, "PoolEth: it has no amount");
 
         _changeStateRewards();
-        uint auxBalance = balances[msg.sender];
+        uint256 auxBalance = balances[msg.sender];
         balances[msg.sender] = 0;
         totalPool -= auxBalance;
 
         payable(msg.sender).transfer(amount);
     }
 
-    function _availableAmount() private view returns (uint) {
+    function _availableAmount() private view returns (uint256) {
         return balances[msg.sender] + getReward();
     }
 
     function _changeStateRewards() private {
-        Reward [] storage rewardsStaker = rewards[msg.sender];
-        for (uint i ; i < rewardsStaker.length; i++) {
-            if(State.DEPOSITED == rewardsStaker[i].state){
+        Reward[] storage rewardsStaker = rewards[msg.sender];
+        for (uint256 i; i < rewardsStaker.length; i++) {
+            if (State.DEPOSITED == rewardsStaker[i].state) {
                 rewardsStaker[i].state = State.WITHDRAWN;
             }
         }
@@ -88,25 +83,25 @@ contract PoolEth is Ownable {
         require(msg.value > 0, "PoolEth: amount can not be 0");
         require(totalPool > 0, "PoolEth: no users in pool");
         deposits.push(Deposit(msg.value, block.timestamp));
-        uint idDeposit = deposits.length -1;
-        for (uint i; i < stakers.length; i++) {
-            if(balances[stakers[i]] > 0){
-                uint percentage = (balances[stakers[i]] * 100) / totalPool;
-                uint amount = percentage * msg.value / 100;
+        uint256 idDeposit = deposits.length - 1;
+        for (uint256 i; i < stakers.length; i++) {
+            if (balances[stakers[i]] > 0) {
+                uint256 percentage = (balances[stakers[i]] * 100) / totalPool;
+                uint256 amount = (percentage * msg.value) / 100;
                 rewards[stakers[i]].push(Reward(idDeposit, amount, percentage, State.DEPOSITED));
             }
         }
     }
-    
-    function getBalance() external view onlyUser returns(uint) {
+
+    function getBalance() external view onlyUser returns (uint256) {
         return balances[msg.sender];
     }
-    
-    function getReward() public view onlyUser returns(uint) {
-        uint rewardAmount;
-        Reward [] memory rewardsStaker = rewards[msg.sender];
-        for (uint i ; i < rewardsStaker.length; i++) {
-            if(State.DEPOSITED == rewardsStaker[i].state){
+
+    function getReward() public view onlyUser returns (uint256) {
+        uint256 rewardAmount;
+        Reward[] memory rewardsStaker = rewards[msg.sender];
+        for (uint256 i; i < rewardsStaker.length; i++) {
+            if (State.DEPOSITED == rewardsStaker[i].state) {
                 rewardAmount += rewardsStaker[i].amount;
             }
         }
